requirements.txt
aws-cdk-lib==2.108.1
constructs>=10.0.0,<11.0.0
urllib3==2.0.0
boto3==1.29.6
requests==2.31.0